requirements.txt
blinker==1.6.2
click==8.1.6
colorama==0.4.6
flask==2.3.2
flask[async]==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==2.3.6
mysql-connector-python==8.2.0
pika==1.3.2
psycopg[binary]==3.1.18
asgiref==3.7.2
